requirements.txt
asgiref==3.7.2
attrs==24.2.0
backports.zoneinfo==0.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==5.3.0
crispy-bootstrap4==2022.1
cycler==0.11.0
diff==0.6.1
diff-match-patch==20241021
Django==3.2.25
django-crispy-forms==2.0
djangorestframework==3.15.1
etelemetry==0.3.1
filelock==3.12.2
fire==0.7.0
fitz==0.0.1.dev2
Flask==2.2.5
Flask-Cors==5.0.0
fonttools==4.38.0
fuzzywuzzy==0.18.0
httplib2==0.22.0
idna==3.10
ImageHash==4.3.1
imageio==2.31.2
importlib-metadata==4.13.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.5
kiwisolver==1.4.5
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mysql-connector==2.2.9
networkx==2.6.3
nibabel==4.0.2
nipype==1.8.6
numpy==1.21.6
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.0
pandas==1.3.5
pathlib==1.0.1
pdf2docx==0.5.8
Pillow==9.5.0
plum-dispatch==1.7.4
prov==2.0.1
pydot==2.0.0
PyMuPDF==1.22.5
pyparsing==3.1.4
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.2
PyWavelets==1.3.0
pyxnat==1.6.2
rdflib==6.3.2
requests==2.31.0
scikit-image==0.19.3
scipy==1.7.3
simplejson==3.19.3
six==1.17.0
spire==0.4.2
Spire.Doc==12.12.0
sqlparse==0.4.4
termcolor==2.3.0
tifffile==2021.11.2
traits==6.3.2
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
zope.interface==6.4.post2
